Py::Int has been removed from recent PyCXX 7.1, convert to Py::Long
It appears to be gone even with compatability settings; convert all
occurrences to Py::Long

+++ src/App/ApplicationPy.cpp
@@ -869,7 +869,7 @@ PyObject *Application::sCheckLinkDepth(PyObject * /*self*/, PyObject *args)
         return NULL;
 
     PY_TRY {
-        return Py::new_reference_to(Py::Int(GetApplication().checkLinkDepth(depth,false)));
+        return Py::new_reference_to(Py::Long(GetApplication().checkLinkDepth(depth,false)));
     }PY_CATCH;
 }
 
@@ -942,7 +942,7 @@ PyObject *Application::sSetActiveTransaction(PyObject * /*self*/, PyObject *args
         return 0;
 
     PY_TRY {
-        Py::Int ret(GetApplication().setActiveTransaction(name,PyObject_IsTrue(persist)));
+        Py::Long ret(GetApplication().setActiveTransaction(name,PyObject_IsTrue(persist)));
         return Py::new_reference_to(ret);
     }PY_CATCH;
 }
@@ -959,7 +959,7 @@ PyObject *Application::sGetActiveTransaction(PyObject * /*self*/, PyObject *args
             Py_Return;
         Py::Tuple ret(2);
         ret.setItem(0,Py::String(name));
-        ret.setItem(1,Py::Int(id));
+        ret.setItem(1,Py::Long(id));
         return Py::new_reference_to(ret);
     }PY_CATCH;
 }
